refactor: +обновлены импорты; +добавлена проверка некорректных импортов в файлах
