+++ .huskyrc.js
@@ -1,5 +1,6 @@
 module.exports = {
   hooks: {
     ...require('@gpn-prototypes/frontend-configs/.huskyrc.js').hooks,
+    'pre-push': 'yarn test -b --passWithNoTests --watchAll=false',
   },
 };
