+++ src/components/ExcelTable/Header/HeaderWithContextMenu.tsx
@@ -1,9 +1,7 @@
 import React, { ReactElement } from 'react';
 import { withContextMenu } from 'components/ExcelTable/ContextMenu';
-import {
-  HEADER_CONTEXT_ID,
-  HeaderRendererProps,
-} from 'components/ExcelTable/types';
+import { ContextMenuId } from 'components/ExcelTable/enums';
+import { HeaderRendererProps } from 'components/ExcelTable/types';
 
 import { DefaultHeader as Header } from './DefaultHeader';
 
@@ -11,6 +9,6 @@ export const HeaderWithContextMenu = (
   props: HeaderRendererProps,
 ): ReactElement =>
   withContextMenu(<Header {...props} />, {
-    id: HEADER_CONTEXT_ID,
+    id: ContextMenuId.HEADER,
     collect: () => ({ idx: props.column.idx }),
   });
