+++ src/components/ExcelTable/Models/index.ts
@@ -1,0 +1,4 @@
+import GridColumnEntity from './GridColumnEntity';
+import OptionEntity from './OptionEntity';
+
+export { GridColumnEntity, OptionEntity };
