+++ src/components/ExcelTable/enums.ts
@@ -1,0 +1,28 @@
+export enum ContextMenuId {
+  HEADER = 'header-context-menu',
+  ROW = 'row-context-menu',
+}
+
+export enum CategoryIcon {
+  LICENSING_ROUND_ICON = 'LICENSING_ROUND_ICON',
+  FIELD_ICON = 'FIELD_ICON',
+  FORMATION_ICON = 'FORMATION_ICON',
+  OIL_POOL_ICON = 'OIL_POOL_ICON',
+  WELL_ICON = 'WELL_ICON',
+}
+
+export enum VisibleKeys {
+  CALCULATION = 'calc',
+  TABLE = 'table',
+  TREE = 'tree',
+}
+
+export enum TableEntities {
+  GEO_CATEGORY = 'RBDomainEntity',
+  CALC_PARAM = 'Attribute',
+  GEO_CATEGORY_TYPE = 'GeoObjectType',
+  RISK = 'Risk',
+  ID = 'ID',
+  SPLITTER = 'Splitter',
+  NONE = 'None',
+}
