+++ src/components/ExcelTable/index.ts
@@ -1,3 +1,8 @@
 import { ExcelTable } from './ExcelTable';
 
-export default ExcelTable;
+export * from './enums';
+export * from './types';
+export * from './utils';
+export * from './Models';
+
+export { ExcelTable };
