+++ src/components/ExcelTable/utils/columns/createColumn.test.ts
@@ -1,4 +1,4 @@
-import { TableEntities } from 'components/ExcelTable/types';
+import { TableEntities } from 'components/ExcelTable/enums';
 
 import { createColumn } from './createColumn';
