+++ src/components/ExcelTable/utils/columns/index.ts
@@ -1,0 +1,5 @@
+import { columnsFactory } from './columnsFactory';
+import { createColumn } from './createColumn';
+import { getSameColumnKeys } from './getSameColumnKeys';
+
+export { getSameColumnKeys, createColumn, columnsFactory };
