+++ src/components/ExcelTable/utils/hasContextMenu.ts
@@ -1,5 +1,5 @@
+import { TableEntities } from 'components/ExcelTable/enums';
 import { isNoneColumnType } from 'components/ExcelTable/helpers';
-import { TableEntities } from 'components/ExcelTable/types';
 
 export function hasContextMenu(type: TableEntities): boolean {
   return !(isNoneColumnType(type) || type === TableEntities.GEO_CATEGORY_TYPE);
